refactor: builder abstract struct
