+++ pkg/builder/devcontainer.go
@@ -18,8 +18,6 @@ import (
 
 	"github.com/daytonaio/daytona/pkg/builder/devcontainer"
 	"github.com/daytonaio/daytona/pkg/docker"
-	"github.com/daytonaio/daytona/pkg/logger"
-	"github.com/daytonaio/daytona/pkg/workspace"
 	"github.com/docker/docker/api/types"
 	"github.com/docker/docker/api/types/container"
 	"github.com/docker/docker/api/types/image"
@@ -33,17 +31,8 @@ type BuildOutcome struct {
 	RemoteWorkspaceFolder string `json:"remoteWorkspaceFolder"`
 }
 
-type DevcontainerBuilderConfig struct {
-	buildId                      string
-	project                      workspace.Project
-	loggerFactory                logger.LoggerFactory
-	localContainerRegistryServer string
-	projectVolumePath            string
-}
-
 type DevcontainerBuilder struct {
-	BuilderPlugin
-	DevcontainerBuilderConfig
+	*Builder
 	buildImageName     string
 	user               string
 	postCreateCommands []string
@@ -88,7 +77,7 @@ func (b *DevcontainerBuilder) CleanUp() error {
 		return err
 	}
 
-	err = cli.ContainerRemove(ctx, b.buildId, container.RemoveOptions{
+	err = cli.ContainerRemove(ctx, b.id, container.RemoveOptions{
 		Force: true,
 	})
 	if err != nil {
@@ -184,7 +173,7 @@ func (b *DevcontainerBuilder) buildDevcontainer() error {
 		}
 	}(&buildOutcome)
 
-	result, err := dockerClient.ExecSync(b.buildId, execConfig, w)
+	result, err := dockerClient.ExecSync(b.id, execConfig, w)
 	if err != nil {
 		return err
 	}
@@ -207,7 +196,7 @@ func (b *DevcontainerBuilder) buildDevcontainer() error {
 
 	//	todo: commit sha from project Repository
 	tag := "latest"
-	imageName := fmt.Sprintf("%s:%s", b.localContainerRegistryServer+"/p-"+b.buildId, tag)
+	imageName := fmt.Sprintf("%s:%s", b.localContainerRegistryServer+"/p-"+b.id, tag)
 
 	_, err = builderCli.ContainerCommit(context.Background(), buildOutcome.ContainerId, container.CommitOptions{
 		Reference: imageName,
@@ -246,7 +235,7 @@ func (b *DevcontainerBuilder) readConfiguration() error {
 		Tty:          true,
 	}
 
-	result, err := dockerClient.ExecSync(b.buildId, execConfig, nil)
+	result, err := dockerClient.ExecSync(b.id, execConfig, nil)
 	if err != nil {
 		return err
 	}
@@ -305,7 +294,7 @@ func (b *DevcontainerBuilder) startContainer() error {
 		return err
 	}
 
-	dir := "/tmp/" + b.buildId + "/var/lib/docker"
+	dir := "/tmp/" + b.id + "/var/lib/docker"
 	err = os.MkdirAll(dir, 0755)
 	if err != nil {
 		return err
@@ -323,12 +312,12 @@ func (b *DevcontainerBuilder) startContainer() error {
 			filepath.Dir(b.getLocalDockerSocket()) + ":/tmp/docker",
 			dir + ":/var/lib/docker",
 		},
-	}, nil, nil, b.buildId)
+	}, nil, nil, b.id)
 	if err != nil {
 		return err
 	}
 
-	if err := cli.ContainerStart(ctx, b.buildId, container.StartOptions{}); err != nil {
+	if err := cli.ContainerStart(ctx, b.id, container.StartOptions{}); err != nil {
 		return err
 	}
 
@@ -349,7 +338,7 @@ func (b *DevcontainerBuilder) startDocker() error {
 		Cmd:          []string{"dockerd", "-H", "unix:///tmp/docker/docker.sock", "-H", "unix:///var/run/docker.sock", "--insecure-registry", b.localContainerRegistryServer},
 	}
 
-	execResp, err := cli.ContainerExecCreate(ctx, b.buildId, execConfig)
+	execResp, err := cli.ContainerExecCreate(ctx, b.id, execConfig)
 	if err != nil {
 		return err
 	}
@@ -380,5 +369,5 @@ func (b *DevcontainerBuilder) getBuilderDockerClient() (*client.Client, error) {
 }
 
 func (b *DevcontainerBuilder) getLocalDockerSocket() string {
-	return filepath.Join("/tmp", b.buildId, "docker", "docker.sock")
+	return filepath.Join("/tmp", b.id, "docker", "docker.sock")
 }
